oelp_backend/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-gis==1.0
django-filter==24.2
# psycopg v2 wheels fail to build on Python 3.13. Use psycopg v3 binary wheel.
psycopg[binary]==3.2.3
django-cors-headers==4.4.0
django-storages==1.14.4
boto3==1.34.162
razorpay==1.4.2
Pillow==10.4.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.7
PyJWT==2.9.0
reportlab==4.2.2
whitenoise==6.7.0
gunicorn
dj-database-url 
setuptools<81
